plaid_django/requirements.txt
amqp==2.4.2
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
Django==2.2.1
django-celery-beat==1.4.0
django-celery-results==1.0.4
django-cors-middleware==1.3.1
django-filter==2.1.0
django-oauth-toolkit==1.2.0
django-timezone-field==3.0
djangorestframework==3.9.4
idna==2.8
kombu==4.5.0
Markdown==3.1
oauthlib==3.0.1
plaid-python==3.1.1
psycopg2==2.8.2
python-crontab==2.3.6
python-dateutil==2.8.0
pytz==2019.1
redis==3.2.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.2
uWSGI==2.0.18
vine==1.3.0
